requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
strawberry-graphql[fastapi]>=0.220.0
httpx>=0.25.0
